+++ frameworks/react-native/src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './Field/index.ts';
 export * from './FieldArray/index.ts';
-export * from './Form/index.ts';